Refactor laser settings and improve hardware setup UI
- Changed `LaserPower` from `float` to `double` for precision.
- Updated `HardWareSetupCommand` to accept a `Window` parameter.
- Set `WindowStartupLocation` to `CenterOwner` and `ResizeMode` to `NoResize` in `HardwareSetupWindow.xaml`.
- Enhanced `NumericUpDown` controls with `DecimalPlaces`, `Minimum`, and `Maximum` properties for better input validation.
- Moved `using` directive for `System` and added `HardwareSetupWindow_Loaded` event handler in `HardwareSetupWindow.xaml.cs`.
- Updated `HardWare Setup` menu item in `MainWindow.xaml` to pass the current window as a command parameter.

+++ DrawingCanvasLib/Utils.cs
@@ -54,7 +54,7 @@ namespace DrawingCanvasLib
         public double FoucsHeight { get; set; } = 0.0;
         public int MarkingPasses { get; set; } = 1;
         public int MarkingSpeed { get; set; } = 1;
-        public float LaserPower { get; set; } = 1;
+        public double LaserPower { get; set; } = 1;
         public int LaserFrequnency { get; set; } = 1;
         public double PulseWidth { get; set; } = 1;
         public double LaserOnDelay { get; set; } = 0;
